Actualizar

+++ backend/src/main/java/io/github/projektalmanac/amoxtli/backend/entity/User.java
@@ -12,7 +12,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
     private String name;
     private String lastName;
     private String email;
@@ -27,5 +27,4 @@ public class User {
     @OneToMany(targetEntity = Book.class,fetch = FetchType.EAGER,cascade = CascadeType.ALL)
     @JoinColumn(name = "id_book")
     private List<Book> books = new ArrayList<>();
-
 }
